css/image-fit/js/getComputedStyle/preserveAspectRatio/xMidYMid_slice.html

<!DOCTYPE html>
<html>
	<head>
		<title>getComputedStyle SVG preserveAspectratio='xMidYMid slice'</title>
		<script src="/resources/testharness.js"></script>
		<script src="/resources/testharnessreport.js"></script>
	</head>
	<body>
		<img id="test" src="xMidYMid_slice.svg">
		<script>
		test(function() {
		  var e = document.getElementById('test');
		  assert_equals(getComputedStyle(e, '').OObjectFit, 'auto');
		  assert_equals(getComputedStyle(e, '').OObjectPosition, '50% 50%');
		});
		</script>
	</body>
</html>
